new save and load setting for UCF, JHMDB

+++ train_dab_conv_trans_.py
@@ -131,7 +131,10 @@ def main_worker(cfg):
     max_accuracy = 0.0  
     for epoch in range(cfg.CONFIG.TRAIN.START_EPOCH, cfg.CONFIG.TRAIN.EPOCH_NUM):
         if epoch > 0:
-            last_epoch = os.path.join(cfg.CONFIG.LOG.BASE_PATH, exp_name, cfg.CONFIG.LOG.SAVE_DIR, f'ckpt_epoch_{epoch-1:02d}.pth')
+            if not cfg.CONFIG.DATA.DATASET_NAME == 'jhmdb':
+                last_epoch = os.path.join(cfg.CONFIG.LOG.BASE_PATH, exp_name, cfg.CONFIG.LOG.SAVE_DIR, f'ckpt_epoch_{epoch-1:02d}.pth')
+            else:
+                last_epoch = os.path.join(cfg.CONFIG.LOG.BASE_PATH, exp_name, cfg.CONFIG.LOG.SAVE_DIR, f'ckpt_epoch_{epoch-1:03d}.pth')
         else: 
             last_epoch = ""
         if os.path.isfile(last_epoch):
@@ -150,11 +153,12 @@ def main_worker(cfg):
         if cfg.DDP_CONFIG.GPU_WORLD_RANK == 0 and (
                 epoch % cfg.CONFIG.LOG.SAVE_FREQ == 0 or epoch == cfg.CONFIG.TRAIN.EPOCH_NUM - 1):
             save_checkpoint(cfg, epoch, model, max_accuracy, optimizer, lr_scheduler, scaler)
-        if len(os.listdir(os.path.join(cfg.CONFIG.LOG.BASE_PATH, exp_name, cfg.CONFIG.LOG.SAVE_DIR))) > 20:
-            epochs_folder = os.listdir(os.path.join(cfg.CONFIG.LOG.BASE_PATH, exp_name, cfg.CONFIG.LOG.SAVE_DIR))
-            epochs_folder.sort()
-            oldest_epoch = epochs_folder[0]
-            os.remove(os.path.join(cfg.CONFIG.LOG.BASE_PATH, exp_name, cfg.CONFIG.LOG.SAVE_DIR, oldest_epoch))
+        if os.path.exists(os.path.join(cfg.CONFIG.LOG.BASE_PATH, exp_name, cfg.CONFIG.LOG.SAVE_DIR)):
+            if len(os.listdir(os.path.join(cfg.CONFIG.LOG.BASE_PATH, exp_name, cfg.CONFIG.LOG.SAVE_DIR))) > 20:
+                epochs_folder = os.listdir(os.path.join(cfg.CONFIG.LOG.BASE_PATH, exp_name, cfg.CONFIG.LOG.SAVE_DIR))
+                epochs_folder.sort()
+                oldest_epoch = epochs_folder[0]
+                os.remove(os.path.join(cfg.CONFIG.LOG.BASE_PATH, exp_name, cfg.CONFIG.LOG.SAVE_DIR, oldest_epoch))
 
         if (epoch % cfg.CONFIG.VAL.FREQ == 0 or epoch == cfg.CONFIG.TRAIN.EPOCH_NUM - 1):
             if cfg.CONFIG.DATA.DATASET_NAME == 'ava':
